UPDATE: chat id fields to big interger

+++ src/entity/event.py
@@ -5,4 +5,4 @@ from entity.base_entity import BaseEntity
 
 class Event(BaseEntity):
     time = peewee.IntegerField(primary_key=True)
-    chat_id = peewee.IntegerField()
+    chat_id = peewee.BigIntegerField()
