+++ src/entity/message.py
@@ -7,6 +7,6 @@ from entity.user import User
 class Message(BaseEntity):
     msg_id = peewee.IntegerField(primary_key=True)
     text = peewee.TextField()
-    chat_id = peewee.IntegerField()
+    chat_id = peewee.BigIntegerField()
     user = peewee.ForeignKeyField(model=User)
     time = peewee.DateTimeField()
